telegram_bot/requirements.txt
requests~=2.32.3
python-telegram-bot~=22.1
redis~=6.1.0